Script for normalizing datasets
The code gives as output the ratio for the normalization, which should be multiplied to the dataset which needs to be normalized. Both reference data (dataset to which the other is normalized to) and target data (dataset which has to be normalized) has to be cropped into a fitting size with the location of matching voxels, before using the code.